requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-yasg==1.21.8
Pillow==11.1.0
django-environ==0.11.2
gunicorn==23.0.0
whitenoise==6.7.0
djoser==2.2.3
psycopg[binary]==3.2.10
